% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowFor.R
\name{snowFor}
\alias{snowFor}
\title{snowFor multithreading map-like function based on snow and foreach}
\usage{
snowFor(
  x,
  FUN,
  pre_fun = NULL,
  varlist = NULL,
  cores = parallel::detectCores(),
  env = globalenv(),
  show_workers_msg = F
)
}
\arguments{
\item{x}{list to loop. Accept vector, list and data.frame(by row)}

\item{FUN}{the function for mapping}

\item{pre_fun}{prepare function to init all nodes}

\item{varlist}{name string vector of objects to be exported to nodes: "a_variable"}

\item{cores}{number of threads}

\item{env}{env to store the cl object. defalut: globalenv()}

\item{show_workers_msg}{show workers' message or not}
}
\value{

}
\description{
Errors won't break the loop.
Instead the loop will return an ThreadError object and keep going.
}
\examples{
go_fun = function(x){
  Sys.sleep(1)
  x
 }
a = snowFor(1:10,go_fun,cores = 2)


}
