% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtProgressBarETA.R
\name{txtProgressBarETA}
\alias{txtProgressBarETA}
\title{txtProgressBarETA}
\usage{
txtProgressBarETA(
  min = 0,
  max = 1,
  initial = 0,
  char = "=",
  width = NA,
  title,
  label,
  file = ""
)
}
\arguments{
\item{min}{min}

\item{max}{max}

\item{initial}{initial}

\item{char}{char}

\item{file}{}
}
\value{
txtProgressBar object
}
\description{
txtProgressBarETA
}
